refactor : command design pattern 을 AdminService에 적용
